refactor(helm): rename files
