+++ libs/shared/util-services/src/index.ts
@@ -1,2 +1,2 @@
-export * from './lib/colors'
-export * from './lib/api'
+export * from './lib/get-service-state-colors'
+export * from './lib/build-edit-service-payload'
